refactor: change prefix from substrate to kusama
