+++ src/kusama/index.ts
@@ -1,0 +1,3 @@
+import { useKusama, KusamaContextProvider } from './KusamaContext'
+
+export { useKusama, KusamaContextProvider }
