--- src/substrate/index.ts
@@ -1,3 +1,0 @@
-import { useSubstrate, SubstrateContextProvider } from './SubstrateContext'
-
-export { useSubstrate, SubstrateContextProvider }
